Use a regular array instead of a set of known IDs
This is just me being a little nitpicky. While I don't have proper
benchmarks, I think using a Set is redundant because converting an array
to a Set could be O(n) anyway, while looking for a value in an array is
on average O(n) (though it could be better because it doesn't have to
traverse the entire array every time).